requirements.txt
click==5.1
MySQL-python==1.2.5
ordereddict==1.1
python-dotenv==0.3.0
python-slugify==1.1.4
PyYAML==3.11
SQLAlchemy==1.0.9
Unidecode==0.4.18
wheel==0.24.0
